+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## development
+- Added F# (#43)
+- Added additional C# file extensions (#43)
 - Added setting to force big IDE icon (#27)
 - Added setting to hide elapsed time (#37)
 - Added additional C++ file extensions (#40)
@@ -7,7 +9,7 @@
 - Fixed file not reappearing after removing project description (#31)
 - Fixed debug folder being created after when debug logging has been disabled
 - Updated Java-DiscordRPC to v2.0.1
-- Updated Java-DiscordRPC to v1.3.6
+- Updated JGroups to 4.0.12
 
 ## v0.8.0 - Project descriptions and even more languages
 - Added the ability to add project descriptions (#24)